Bootstrap aot module, copy testing infrastructure from runtime.
