+++ aot/README.md
@@ -1,0 +1,3 @@
+# Chicory AOT
+
+Experimental Wasm -> JVM bytecode AOT engine.
