+++ pom.xml
@@ -43,6 +43,7 @@
     <module>wasi</module>
     <module>wasm-corpus</module>
     <module>cli</module>
+    <module>aot</module>
   </modules>
 
   <scm>
@@ -121,6 +122,11 @@
       <dependency>
         <groupId>com.dylibso.chicory</groupId>
         <artifactId>runtime</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>com.dylibso.chicory</groupId>
+        <artifactId>aot</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
