+++ test-gen-plugin/src/main/java/com/dylibso/chicory/maven/TestGenMojo.java
@@ -117,6 +117,9 @@ public class TestGenMojo extends AbstractMojo {
     @Parameter(defaultValue = "[]")
     private List<String> excludedWasts;
 
+    @Parameter(required = false, defaultValue = "false")
+    private boolean useAot;
+
     /**
      * The current Maven project.
      */
@@ -167,7 +170,8 @@ public class TestGenMojo extends AbstractMojo {
                         sourceDestinationFolder,
                         excludedTests,
                         excludedMalformedWasts,
-                        excludedInvalidWasts);
+                        excludedInvalidWasts,
+                        useAot);
 
         // Create destination folders
         if (!compiledWastTargetFolder.mkdirs()) {
